requirements.txt
fastapi==0.115.5
starlette==0.41.3
uvicorn==0.32.0
python-dotenv==1.0.0
sqlalchemy==2.0.36
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
email-validator==2.3.0
psycopg2-binary==2.9.11
redis==5.0.1
slowapi==0.1.9
alembic==1.13.1